Shop refactor pt3
Parameterized xml reading.
Created a list of all materials, runes, and items to keep order.
Created a dictionary of all items for easy lookup.
That dictionary is now being used to display the price on the left
panel.

TODO : Correctly switch between player and shop inventory to be
displayed on the right.