Refactored InventoryButtons
Inventory Buttons now derive their functionality from the left-page of
the book. The left-page UI scripts will define a function for how
buttons should behave and pass that to the right-page of the book that
acts as the spawner and manager for all Inventory Buttons. The
right-page will correctly update and refresh the behavior of old and new
buttons to reflect how the left-page wants those buttons to behave.