Create an order board that generates random orders
Created a system that can generate random orders - creates x number of
orders where x is the current day number.
Created some UI that can be accessed by clicking the board on the
counter of the storefront.
Orders are like sticky notes and can be moved around with the player's
discretion.