Created a scene manager for scene transitions
The scene manager attaches to the Master GM, so simply call
MasterGameManager.instance.sceneManager to access it.
Tried to make a fade to black/clear script, but it get's a little wonky
as it delays the scene load - allowing the player to do other things
while it's fading - so it's commented for now.