Added a Fade to scene transitions
Include a Canvas with a "Fader" prefab found in UI->ScreenFade in order
to have this functionality.
You will also need an instance of the MasterGameManager in your scene
for this to work. So add a GameManagerLoader prefab from the GameManager
folder, or just access your game from the Workshop