Added logic to recipe UI
Can no longer add ingredients if not enough in stock.
When adding ingredients, decrements the stock accordingly.
Can remove ingredients from the left panel, this refills the stock
(because you haven't pinned yet).
Created two pin buttons that become interactable once the recipe is met,
and deactivate when the recipe is incomplete.