Created logic for the Workboard
There is now a WorkboardManager script attached to the Master Game
Manager.
It contains a list of work orders which represent the items the player
is currently working on.
The RecipeUI properly "pins" to the workboard.
The WorkOrder has attributes like: score, stage, minigames played so far
to make it, if random mode is on for that order, and the number it is in
the WorkboardManager's list
Each WorkOrder has an UpdateOrder function that will (presumably) get
called during the results screen of a minigame to make progress towards
that order.