Updated UI Sounds
Updated UI sounds to be ones that I made instead of ones I got from the
internet. Also, I found a bug where the sounds keep playing if you click
buy, even if you don't buy the item. This has to do with the way the
TradeManager works with the ShopManager, so I'll have to fix this later.
It may require completely revamping the AudioManager at the worst case.